requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
tkinterdnd2==0.4.3
urllib3==2.3.0

pyinstaller~=6.12.0